+++ uncertainties/core.py
@@ -445,6 +445,12 @@ class AffineScalarFunc(object):
 
         This mapping is cached, for subsequent calls.
         """
+        warn(
+            f"{self.__class__.__name__}.derivatives() is deprecated. It will "
+            f"be removed in a future release.",
+            FutureWarning,
+            stacklevel=2,
+        )
 
         if not self._linear_part.expanded():
             self._linear_part.expand()
@@ -469,6 +475,15 @@ class AffineScalarFunc(object):
         object take scalar values (and are not a tuple, like what
         math.frexp() returns, for instance).
         """
+        warn(
+            f"{self.__class__.__name__}.error_components() is currently an "
+            f"instance method. This method is deprecated. In a future release it will "
+            f"be replaced with an instance property by the same name. It will be "
+            f"accessed by {self.__class__.__name__}.error_components (with no "
+            f"parentheses).",
+            FutureWarning,
+            stacklevel=2,
+        )
 
         # Calculation of the variance:
         error_components = {}